input validation, word counter refactor

+++ src/main/java/app/WordCounter.java
@@ -2,6 +2,6 @@ package app;
 
 public interface WordCounter {
 
-    int countWords(String inputString);
+    long countWords(String inputString);
 
 }
