+++ src/main/java/output/ConsoleWriter.java
@@ -2,7 +2,7 @@ package output;
 
 public class ConsoleWriter {
 
-    public void writeWordCountOutput(int count) {
+    public void writeWordCountOutput(long count) {
         System.out.println("Number of words: " + count);
     }
